requirements.txt
requests==2.28.1
bs4==0.0.1
schedule==1.1.0
pyTelegramBotAPI==4.8.0
psycopg2-binary==2.9.5
beautifulsoup4==4.11.1
numpy~=1.24.0
matplotlib~=3.6.2
boto3~=1.26.35